% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROpenCV.R
\docType{package}
\name{ROpenCV-package}
\alias{ROpenCV}
\alias{ROpenCV-package}
\title{ROpenCV: A package to install OpenCV within R}
\description{
This is a utility package that installs OpenCV within R for use
 by other packages.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

